requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
astor==0.7.1
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.5
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==2.0.0
multidict==4.5.2
numpy==1.16.2
pbr==5.1.3
protobuf==3.7.1
python-dateutil==2.8.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
websockets==6.0
Werkzeug==0.15.2
yarl==1.3.0
